OD_PROJEKT/requirements.txt
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-Limiter==3.5.0
Flask-Login==0.6.3
Flask-Migrate==4.0.5
flask-msearch==0.2.9.2
Flask-SQLAlchemy==3.1.1
Flask-Uploads==0.2.1
Flask-WhooshAlchemy==0.56
Flask-WTF==1.2.1
Jinja2==3.1.2
limits==3.7.0
SQLAlchemy==2.0.23
Werkzeug==3.0.1
WTForms==3.1.1
zxcvbn==4.4.28
pycryptodome==3.12.0